Re group all new componets
