+++ libs/core/src/components/content/divider/divider.style.css.ts
