+++ libs/core/src/components/content/divider/index.ts
