+++ libs/core/src/components/content/spacer/spacer.ts
@@ -1,10 +1,10 @@
 import {
   gdsCustomElement,
   html,
-} from '../../utils/helpers/custom-element-scoping'
-import { GdsElement } from '../../gds-element'
-import { tokens } from '../../tokens.style'
-import { styleExpressionProperty } from '../../utils/decorators/style-expression-property'
+} from '../../../utils/helpers/custom-element-scoping'
+import { GdsElement } from '../../../gds-element'
+import { tokens } from '../../../tokens.style'
+import { styleExpressionProperty } from '../../../utils/decorators/style-expression-property'
 
 @gdsCustomElement('gds-spacer')
 export class GdsSpacer extends GdsElement {
