+++ libs/core/src/components/content/text/index.ts
