+++ libs/core/src/components/content/text/text.style.css.ts
