+++ libs/core/src/components/content/text/text.test.ts
@@ -1,6 +1,6 @@
 import { expect, fixture, html as testingHtml } from '@open-wc/testing'
 import '@sebgroup/green-core/components/text'
-import type { GdsCard } from '@sebgroup/green-core/components/text'
+import type { GdsText } from '@sebgroup/green-core/components/text'
 
 import { htmlTemplateTagFactory } from '@sebgroup/green-core/scoping'
