+++ libs/core/src/components/index.ts
@@ -6,15 +6,15 @@ export * from './menu-button/menu-button'
 export * from './context-menu/context-menu'
 export * from './datepicker/datepicker'
 export * from './dropdown/dropdown'
-export * from './grid/grid'
-export * from './card/card'
-export * from './flex/flex'
-export * from './container/container'
-export * from './img/img'
-export * from './video/video'
-export * from './text/text'
-export * from './divider/divider'
-export * from './spacer/spacer'
+export * from './layout/grid/grid'
+export * from './layout/card/card'
+export * from './layout/flex/flex'
+export * from './layout/container/container'
+export * from './media/img/img'
+export * from './media/video/video'
+export * from './content/text/text'
+export * from './content/divider/divider'
+export * from './content/spacer/spacer'
 export * from './grouped-list/grouped-list'
 export * from './icon/icon'
 export * from './segmented-control/segmented-control'
