+++ libs/core/src/components/layout/card/card.stories.ts
@@ -1,13 +1,13 @@
 import type { Meta, StoryObj } from '@storybook/web-components'
 import './card'
-import '../img'
-import '../video'
-import '../container'
-import '../text'
-import '../divider'
-import '../icon/icons/arrow-down.js'
-import '../icon/icons/arrow-right.js'
-import '../icon/icons/circles-three.js'
+import '../../media/img'
+import '../../media/video'
+import '../../layout/container'
+import '../../content/text'
+import '../../content/divider'
+import '../../icon/icons/arrow-down.js'
+import '../../icon/icons/arrow-right.js'
+import '../../icon/icons/circles-three.js'
 import { html } from 'lit'
 
 /**
