+++ libs/core/src/components/layout/card/card.style.css.ts
