+++ libs/core/src/components/layout/container/container.ts
@@ -1,11 +1,11 @@
 import {
   gdsCustomElement,
   html,
-} from '../../utils/helpers/custom-element-scoping'
-import { GdsElement } from '../../gds-element'
+} from '../../../utils/helpers/custom-element-scoping'
+import { GdsElement } from '../../../gds-element'
 import { GdsFlex } from '../flex'
-import { tokens } from '../../tokens.style'
-import { styleExpressionProperty } from '../../utils/decorators/style-expression-property'
+import { tokens } from '../../../tokens.style'
+import { styleExpressionProperty } from '../../../utils/decorators/style-expression-property'
 
 import ContainerCSS from './container.style.css'
