+++ libs/core/src/components/layout/flex/flex.stories.ts
@@ -1,16 +1,9 @@
 import type { Meta, StoryObj } from '@storybook/web-components'
-import '../card'
 import { html } from 'lit'
 
-/**
- * The `gds-container` component is a container element that can be used to wrap other components.
- *
- * @extends `gds-flex`
- *
- */
 const meta: Meta = {
-  title: 'Docs/Layout/Container',
-  component: 'gds-container',
+  title: 'Docs/Layout/Flex',
+  component: 'gds-flex',
   tags: ['autodocs'],
 }
 
@@ -26,12 +19,8 @@ const DefaultParams: Story = {
   },
 }
 
-/**
- * Container
- */
-
-export const Card: Story = {
+export const Flex: Story = {
   ...DefaultParams,
-  name: 'Container',
-  render: (args) => html` <gds-container> Container </gds-container>`,
+  name: 'Flex',
+  render: (args) => html` <gds-flex> Flex </gds-flex>`,
 }
