+++ libs/core/src/components/layout/flex/flex.ts
@@ -1,10 +1,10 @@
 import {
   gdsCustomElement,
   html,
-} from '../../utils/helpers/custom-element-scoping'
-import { GdsElement } from '../../gds-element'
-import { tokens } from '../../tokens.style'
-import { styleExpressionProperty } from '../../utils/decorators/style-expression-property'
+} from '../../../utils/helpers/custom-element-scoping'
+import { GdsElement } from '../../../gds-element'
+import { tokens } from '../../../tokens.style'
+import { styleExpressionProperty } from '../../../utils/decorators/style-expression-property'
 
 import FlexCSS from './flex.style.css'
