+++ libs/core/src/components/media/img/img.stories.ts
