+++ libs/core/src/components/media/img/img.style.css.ts
