+++ libs/core/src/components/media/img/index.ts
