+++ libs/core/src/components/media/video/video.style.css.ts
